requirements.txt
Django
python-dotenv
psycopg
gunicorn